Test cmake install and build examples (#578)
* [github] test install with cmake and build examples

* [cmake] install custom cmake find modules

* [cmake] add required dependencies in exported config

* [cmake] only use custom FindHDF5 from config if cmake < 3.19

* [clang] remove deprecated implicit definitions

* [examples] require c++17 if not using boost

* [cmake] hdf5 link must be public

* [examples] remove complex_io.cpp (linking error)

* [examples] fix image_h5 example

* [examples] fix writer examples

* [doc] add github workflow badge to readme

* [cmake] remove unecessary addition to module path

* Revert "[examples] remove complex_io.cpp (linking error)"

This reverts commit 23cd78ca5828e1f347a205365ded52d94cbb9e98.

* [clang] resolve compiler warnings

* [cmake] add option to include the examples in the project

* [examples] fix complex_io example

* [clang] resolve compiler warnings in examples

+++ examples/std_vector_io.cpp
@@ -50,7 +50,7 @@ int main()
   //
   // retrieving the data back from disk
   //
-  DataType read_data(dataset.dataspace().size()); //allocate enough memory
+  DataType read_data(static_cast<size_t>(dataset.dataspace().size())); //allocate enough memory
   dataset.read(read_data);
 
   //
