+++ test/node/link_target_test.cpp
@@ -36,7 +36,7 @@
 namespace { 
   static size_t get_path_size(const fs::path& path) {
   #if BOOST_VERSION < 106000
-    return std::distance(path.begin(), path.end());
+    return static_cast<size_t>(std::distance(path.begin(), path.end()));
   #else
     return path.size();
   #endif
